refactor: second style pass over the fast path (own review)
Hand-review against the original brief (no magic values, no raw offset
grabs, declarative over imperative), beyond the agent findings:

- name the fast-lane caps: depth > 8 -> MAX_FAST_LANE_NESTING,
  n > 4096 -> MAX_SUBSTITUTIONS
- SID descriptor rule: destructure the payload with a slice pattern
  (&[_revision, sub_count, ..]) instead of peeking data[off + 1]; same
  for write_sid, where one pattern replaces the len<8 check, two indexed
  peeks, and a manual big-endian fold (u64::from_be_bytes)
- ends_stream: one slice-pattern rule (Some([]) | Some([EOF, ..]))
  replaces three hand-rolled consumed-vs-len trailer checks
- raw_next: compare record-slice bounds in u64 domain (the usize cast
  could truncate the guard on 32-bit targets)
- module doc still claimed bails route to "the existing render-direct
  path" - rewritten for the single-walker architecture; import token/
  single_instance_offset instead of six full-path references

Sweep byte-identical; 99 tests; bench unchanged (52-53ms M3 Pro).

+++ src/evtx_chunk.rs
@@ -433,7 +433,7 @@ fn raw_next<'a>(
 
     let binxml_start = record_start + EVTX_RECORD_HEADER_SIZE as u64;
     let binxml_end = binxml_start.saturating_add(binxml_data_size as u64);
-    if binxml_end as usize > chunk.data.len() {
+    if binxml_end > chunk.data.len() as u64 {
         walk.exhausted = true;
         return Some(Err(EvtxError::FailedToParseRecord {
             record_id: record_header.event_record_id,
